Changes in balance check

+++ mb-xrpl/app.js
@@ -65,7 +65,7 @@ async function main() {
                 await new Setup().checkBalance();
             }
             else if (process.argv.length >= 4 && process.argv[2] === 'wait-for-funds') {
-                await new Setup().waitForFunds(process.argv[3], parseInt(process.argv[4]));
+                await new Setup().waitForFunds(process.argv[3], parseFloat(process.argv[4]));
             }
             else if (process.argv.length >= 3 && process.argv[2] === 'prepare') {
                 await new Setup().prepareHostAccount(process.argv[3]);
